Clean-up: Quiet macOS/BSD compiler warnings, update CI/CD

See merge request dps8m/dps8m!193

+++ src/dps8/libtelnet.c
@@ -436,7 +436,7 @@ static int _environ_telnet(telnet_t *telnet, unsigned char type,
                         (unsigned)buffer[0] != TELNET_ENVIRON_IS &&
                         (unsigned)buffer[0] != TELNET_ENVIRON_INFO) {
                 _error(telnet, __LINE__, __func__, TELNET_EPROTOCOL, 0,
-                                "telopt %d subneg has invalid command", type);
+                                "telopt %ld subneg has invalid command", (long) type);
                 return 0;
         }
